refac: Move storage column name classes to better locations
