+++ source/databricks/calculation_engine/package/databases/wholesale_results_internal/wholesale_result_column_names.py
@@ -16,18 +16,22 @@
 from .result_column_names import ResultColumnNames
 
 
-class EnergyResultColumnNames(ResultColumnNames):
-    time_series_type = "time_series_type"
-    neighbor_grid_area_code = "neighbor_grid_area_code"
-    """The delta table column name ought to be renamed to 'from_grid_area'.
-    It is, however, rather complicated compared to the value unless using Delta Lake column mapping,
-    which comes with its own set of problems/limitations."""
-    balance_responsible_id = "balance_responsible_id"
-    """Obsolete"""
-    aggregation_level = "aggregation_level"
-    """Obsolete"""
+class WholesaleResultColumnNames(ResultColumnNames):
+    """Column names for the wholesale result storage model."""
+
+    amount_type = "amount_type"
+    quantity_unit = "quantity_unit"
+    resolution = "resolution"
+
     metering_point_type = "metering_point_type"
-    metering_point_id = "metering_point_id"
+    settlement_method = "settlement_method"
+    price = "price"
+    amount = "amount"
+    is_tax = "is_tax"
+
+    charge_code = "charge_code"
+    charge_type = "charge_type"
+    charge_owner_id = "charge_owner_id"
+
     quantity = "quantity"
     quantity_qualities = "quantity_qualities"
-    resolution = "resolution"
